undo: clean up bin/hex
Standardizes some bin/hex in hg undo.  hg stores node data in hex form to make
it easier to read its files and because it seems to resolve some parsing issues.
While converting everything to binary seems prudent, it proved non-trivial and
isn't a pressing concern.  If we want to make this change, we should do it
before hg undo is rolled out generally.  As no other extension should directly
access the data it doesn't seem that important.  Interesting information should
be available with the revsets already.

Differential Revision: https://phab.mercurial-scm.org/D188