add additional info to the hg_import_helper "started" response
Summary:
Update the hg_import_helper.py script to include additional information in the
initial CMD_STARTED response that it sends after it has successfully
initialized.

This response now includes:
- A protocol version number, to help catch errors if edenfs somehow ends up
  invoking an incompatible version of the hg_import_helper.py script.
- Whether treemanifest is supported in this repository.
- If treemanifest is supported, the list of treemanifest pack directories.

This eliminates the need for the separate GET_CACHE_PATH command that was
previously sent immediately after the CMD_STARTED response.

This also fixes the code to always let the python code select the pack
directories.  Previously the C++ code assumed the local pack directory was
aways located at ".hg/store/packs/manifest" under the repoPath.  This may not
be correct if repoPath happens to be a repository created with the share
extension.

Reviewed By: wez

Differential Revision: D5560787

fbshipit-source-id: e796222f42927a85e886227b3a7b2ccf9c1ef1bd