treemanifest: move stores to manifestlog
Summary:
This adds the data and history stores to the manifestlog, instead of storing
them on the opener. It also moves construction of the stores to manifestlog
creation time, instead of repo creation time. This is important for bundlerepo
support, since the bundlerepo needs to access the manifestlog as part of the
initial repo setup, and previously the stores weren't added to the opener until
after repo setup.

This patch does not remove the stores from the opener yet. Future patches will
migrate other extensions that depend on the datastores, and then we'll delete
them from the opener.

Test Plan: Ran the tests

Reviewers: #fbhgext, ryanmce

Reviewed By: #fbhgext, ryanmce

Differential Revision: https://phab.mercurial-scm.org/D234