hgsql: make it possible to sqlstrip corrupt commits
Summary:
@davidsp had some issues with the Perforce importer where it created
corrupt commits that cannot be read from the revlog. I had to hand-edit things
to remove these commits from hgsql, as sqlstrip would always try to create a
backup bundle.

I manually patched hgsql to not create backups, and was able to use that to
remove when the issue recurred; provide a more official way to handle this.

Test Plan: Add new test case to test-strip.t and make it pass

Reviewers: durham, rmcelroy

Subscribers: davidsp, #sourcecontrol

Differential Revision: https://phabricator.intern.facebook.com/D5649385