split: only do checkunfinished if rebase is needed
It's a valid use-case to run split inside a histedit plan. The
checkunfinished was added by D161, which was caused by a user report that
the split changes got lost because of rebase failure. If there is no need to
rebase, then checkunfinished is not necessary.

In the future we might want a `splitstate` to make it harder to lose split
commits.

Differential Revision: https://phab.mercurial-scm.org/D488