turn treemanifest back off by default again
Summary:
Currently treemanifest import fails when a mercurial transaction is in progress
(if the tree in question was created by the pending transaction), and this
breaks many mercurial workflows.

Turning treemanifest back off by default, to fall back to the slower but
functional flat manifest import.

D5685880 adds framework to run the eden integration tests with treemanifest
enabled; those tests can be used to tell when treemanifest is working well
enough that we can turn it back on.

Reviewed By: bolinfest

Differential Revision: D5692894

fbshipit-source-id: 5ee84cf73db4cd87bbdaae1edd92c74058fa00e2