metaconfig_types: add as_bytes to PathElement
Summary:
PathElement is a blackbox now and it is hard to read inside it.
as_bytes() should at least enable reading the content of it and allow an attempt to create a string out of it. Alternative would be to create a Path using Path::join(&[PathElement]), but that is a bit too much.

Reviewed By: jsgf

Differential Revision: D5688868

fbshipit-source-id: f30879251e3da7e6b90c70bfbadb901a3d818dd7