refresh data pack store on tree miss
Summary:
this is required together with D5711177 to successfully perform an `hg
amend`.   The changes in D5711177 cause the pending pack files to get written
out and the amend command will then call into eden to change the parents of the
commit.

We need to be able to resolve the tree from the packfiles when this happens,
but since this happens within the default refresh interval in the store code
(which is ~100ms) we need to explicitly refresh the set of pack files.

This is most easily accomplished by forcing a refresh on a tree miss.
This is probably fine if you assume that we won't legitimately be asked
to resolve non-existent trees very frequently.

Reviewed By: simpkins

Differential Revision: D5712623

fbshipit-source-id: 4d0034affcc276f1ae29caac36aa5596e52cd746