fastmanifest: flush data when writepending is called
Summary:
This is required so that eden is able to see the pending trees using the
HG_PENDING mechanism when eg: `hg amend` is used.  The eden dirstate
`setparents` implementation invokes `writepending` so this is part is
pretty straightforward.

What was a bit more difficult to trace through and understand was that
`hg amend` invokes setparents more than once in my simple
manual test scenario (a commit off master, plus a change to a file not
in that existing commit).

Since the pending callback is removed after the first `writepending` call
the subsequent call(s) to `setparents` had a NOP `writepending` and then
eden failed to resolve the data.

Re-establishing the callback after writing the data seems to do the job.

Test Plan: manual testing with `hg amend --debug` in an eden repo

Reviewers: simpkins, mbolin, durham

Reviewed By: durham

Subscribers: medson, mjpieters, net-systems-diffs@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D5711177

Signature: t1:5711177:1503789059:04c06ffb6e7006a5188d5426ae706dd73cd4b150