rust: asyncmemo: allow fill function to be recursive
Summary:
Pass a reference to the cache to the fill function. This allows the
function to be recursive based on memoized values.

This also required quite a bit of restructuring to make sure that locks
and ownership are handled properly during recursive calls. Specifically,
a new `Slot` state - `Polling` - is used to indicate when a thread/task
is currently calling `.poll()` on the future. This contains a list of
futures Tasks which are interested in the state of the slot which can be
notified when it changes state.

Also removed unused Entry API code.

Reviewed By: sid0

Differential Revision: D5652704

fbshipit-source-id: 29cd3fe37d4eb9316235872b7e2e228bf10a016f