treemanifest: implement changegroup file discovery when tree-only
When building a changegroup, the logic relies on processing each manifest to
discover what files to send. In the new tree-only world we don't include the
trees as part of the changegroup, so we don't process the manifests anymore. So
we need to iterate over the tree manually to figure out what files to send.