remotefilelog: make pack cleanup more robust
Summary:
We're seeing exceptions from the unlink cleanup logic hide the actual exception
from the code. In order to debug the actual exception, we need to eat the
cleanup exceptions and in any place where we eat an exception inside of a catch
block, we need to manually rethrow the original exception, otherwise we end up
throwing the second caught exception (from the cleanup logic).

Test Plan:
Ran the tests. This is a particular odd edge case, so I need to run
these bits in the production environment where it was occuring before I can
understand why it was happening.

Reviewers: #fbhgext, ms2316

Reviewed By: ms2316

Subscribers: ms2316

Differential Revision: https://phab.mercurial-scm.org/D586