amend: fix test-treemanifest-treeonly.t after amend's redundant commit removal
Summary: cmdutil.amend was changed recently to get rid of the redundant commit
during the amend operation. As a result, we need to update this test as it was
aware of the redundant commit. This commit achieves the same.

Test Plan: Executed the test after the changes.

Reviewers: #fbhgext, durham, quark

Reviewed By: #fbhgext, quark

Differential Revision: https://phab.mercurial-scm.org/D661

message empty to abort commit. HG: -- HG: user: Saurabh Singh <singhsrb@fb.com>